media-gfx/sane-backends: remove confusing and outdated seds
 * SED: the following did not cause any changes
 *     sed -i -e "/by sane-desc 3.5 from sane-backends/s:sane-backends .*:sane-backends ${ver}:" testsuite/tools/data/html* || die;
 * no-op: -e /by sane-desc 3.5 from sane-backends/s:sane-backends .*:sane-backends 1.2.1:

This is seemingly redundant since upstream ticket:
https://gitlab.com/sane-project/backends/-/issues/228

The sources are changed with a minor version bump to sane-desc (not
sane-backends!) such that the sed no longer applies. Seds are wonderful!
Especially at silently failing and never being noticed.

The tests do still pass, regardless of commented claims in the ebuild.
Is that because the issue was misdiagnosed? Is that because the upstream
ticket to remove timestamps made the issue obsolete? I don't know, and
I'm not fussed about finding out. The tests have passed for years
without this, so that's good enough for me.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>