sys-apps/the_silver_searcher: fix the no-LFS64 patch
According to the documentation glibc still defaults to 32-bit file
offsets on 32-bit arches, therefore tell autoconf to check if it needs
to explicitly set this to 64.

Straight-to-stable revbump on the off chance someone has emerged the
ebuild using the previous version of the patch on a 32-bit system. And
that's enough of messing with this without the usual ~arch revisions.

Signed-off-by: Marek Szuba <marecki@gentoo.org>