media-libs/kvazaar: don't manually enable -Werror
autotools is genius software. When you add a flag called
--disable-werror, it isn't actually a flag called --disable-werror. It's
a flag called AC_ARG_ENABLE(werror), which expands "code if argument
was passed on the command line" and "code if argument was not passed on
the command line" blocks. When you say "--disable-werror" all you
actually do is add a formatted help text, which is otherwise totally
ignored by autoconf.

kvazaar had an AC_ARG_ENABLE that defined an
action-if-argument-was-defined-on-the-command-line of nothing, and an
action-if-argument-was-excluded-on-the-command-line of adding -Werror to
CFLAGS.

Back in commit 019c44a86ad017ea54703b7594b076cb0b840707, this was then
used to work around the default addition of -Werror. Logical.

In the very next release, upstream changed the help text to advertise
the --enable-* form, and switched it to
action-if-argument-was-defined-on-the-command-line of adding -Werror,
action-if-argument-was-excluded-on-the-command-line of nothing.

See:
https://github.com/ultravideo/kvazaar/commit/033bc6bc45b0c6b200a4227208d2fa6263e09166

This naturally meant that passing --disable-werror would add -Werror, in
addition to setting the ignored shell variable enable_werror=no. So for
seven years, we've been building with -Werror. Yay!

During that very version bump, --disable-werror was never removed.
Instead, append-flags was used to add an additional gcc switch back off,
per the redundancy requirements of the Department of Redundancy Department's
required approach to redundantly specifying redundant flags.

This also appears to tickle edge cases in LTO due to libtool dropping
some, but not all flags... sometimes. Leading to LTO being performed
with global -Werror.

Bug: https://bugs.gentoo.org/618434
Closes: https://bugs.gentoo.org/907263
Closes: https://bugs.gentoo.org/924296
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>