media-sound/cdparanoia: EAPI 8, add pkg-config file
I'm porting media-sound/grip's ebuild to meson and it uses pkg-config to
look for cdparanoia. Fedora have the patch I'm adding here to add a .pc file
and it looks like it is upstream, just not in a release.

Unfortunately, cdparanoia doesn't seem to have been migrated to xiph's gitlab.

Signed-off-by: Sam James <sam@gentoo.org>