app-arch/libarchive: enable zstd IUSE by default
23.0 profiles no longer enable USE=zstd globally, but on libarchive
seems it makes very little sense to be off by default. Popular format
that you'd expect libarchive to be able to handle out-of-the-box and
zstd will be pulled by portage either way (no extra dependencies).

But the more immediate motivation is that qtbase[zstd] requires cmake
with zstd support, which is tied with it being enabled on libarchive.
And then recent developments led to force enabling zstd on qtbase.
This will spare users from having to set it themselves.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>