media-gfx/openvdb: 11.0.0 fix abi-compat USE
The values for `-DOPENVDB_USE_*_ABI` are derived from the initial value of `version`.
By changing the value of `version` depending on the abi*-compat USE-flag they are instead derived from the modified value for any USE="!abi11-compat".
That results in `usex abi*-compat`-queries on values outside of the supported +1/-2 range.

Closes: https://bugs.gentoo.org/929926
Signed-off-by: Paul Zander <negril.nx+gentoo@gmail.com>