sys-libs/libseccomp: rebase experimental-loong patches
The fix backported in d58730757b79ce429d5fbe8f35dede6074d4c243 was
already in the loong tree which makes the patch fail to apply.  Rebase
this and also now rather than having a huge patch that doesn't
correspond to VCS with a bunch of checked-in autotools stuff, use the
github diff and eautoreconf for 9999 and loong.

Fixes: d58730757b79 ("sys-libs/libseccomp: backport fix for arch-syscall-check")
Bug: https://bugs.gentoo.org/926648
Signed-off-by: Matoro Mahri <matoro_gentoo@matoro.tk>
Signed-off-by: WANG Xuerui <xen0n@gentoo.org>