net-analyzer/greenbone-feed-sync: drop config file patching in src_test
It turns out that, first, this is not necessary. At least I do not get
sandbox violations in src_test(). Second, it was broken anyways, as it
used $TARGET which is not defined in src_test() (only in python_test()).

Also drop the "${WORKDIR}/${P}" part used in the sed, simply exploit the
fact that we are in "${WORKDIR}/${P}" at this moment and use a relative
path.

Signed-off-by: Florian Schmaus <flow@gentoo.org>