acct-user/redict: new package, add 0
While redict is meant to be a drop-in replacement for redis, it can run
independetly of it or any other fork, with it's own paths for configs
and state.

And while all current forks are backwards compatible, they're not
guaranteed to be forward compatible, while could lead to unexpected
behaviour. Different users and groups for each fork minimize the space
for data corruption.

Signed-off-by: Anna (navi) Figueiredo Gomes <navi@vlhl.dev>
Bug: https://bugs.gentoo.org/928119
Signed-off-by: Petr Vaněk <arkamar@gentoo.org>