x11-drivers/nvidia-drivers: extend 390 postinst to mention kernel
Message does come a bit late (after potential build failures), but
linux-mod-r1 already warns that it's not supported. This just an
extra note about upstream end of support, patches, and just so do
not decide to switch to a newer kernel on a whim.

Users are of course free to use /etc/portage/patches if they know
what they're doing. So there's no hard blocker for kernel version.

Bug: https://bugs.gentoo.org/930208
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>