dev-qt/qtwebchannel: adjust version check for qtdeclarative (qt6)
To recap, this checks for Quick, and then implicitly uses Qml in
a different file. Normally would disable Qt6Qml to disable both
but that gives an "unused option" warning if a *matching version*
qtdeclarative does not exist (and tinderbox automated report).

However, noticed that qtwebchannel-6.7.0 installed some Quick files
and that's because upgraded from 6.7.0_rc2 to 6.7.0 and internally
that's both 6.7.0.

This also shouldn't be using >=, but rather =$(ver_cut 1-3)*.

It'd be simpler+safer to just ignore the unused option warning and
always pass it though (it'd also work for 9999 ebuilds more properly),
but it's not a big enough deal to be worth the noise.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>