sys-libs/glibc: fix CPP definition
On sparc, this might break if e.g. -mcpu is in CFLAGS. It's similar to
the arm case already mentioned.

Reported by Dakon:
"ok, it greps for __sparc_v8 or __sparc_v9 cpp output, but it does not pass the -mcpu
I have in CFLAGS, so it falls back to the compiler default which still is v7"

Fixes: 30e32d9ed408fd786e9c1e16063c1228d123ebc1
Fixes: 39ba3e621469464a59dc7a37e3c41366d7856066
Signed-off-by: Sam James <sam@gentoo.org>