gui-apps/ags: new auth USE flag, cleanup, don't install node_modules to bin dir
This commit adds a new USE flag for authentication which when enabled
adds PAM as a dependency and creates a pam config file for ags.

It also cleans up some of the code by removing src_compile since it is
the same as the default, and also stops manually passing in a libdir.

Finally, now it doesn't copy node_modules into the output directory
since this was unnecessary.

Signed-off-by: DerpDays <34582078+DerpDays@users.noreply.github.com>