x11-drivers/nvidia-drivers: stabilize 550.78 for amd64
Bit early but it's almost identical to 550.76 beside fixing
some Xid error with some laptops, so may as well stable this
one rather htan 550.76.

Note that it does not fix issues with freezing that some laptops
been having (esp. lenovo), that is currently being looked into by
NVIDIA. Keep using the 535 branch if you are affected.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>