x11-plugins/wmcalendar: add patches to fix horrible missing function decls
We kill two birds with one stone. It fails the c99 porting tracker and
will not compile with Modern C compilers such as gcc 14 or clang 16, and
it ALSO fixes LTO errors since -Werror=lto-type-mismatch complained when
the implicit function declaration's assumed type was incorrect and
didn't match the type of the actual live instance of the function.

While we are at it, fix some additional (memory access) errors
discovered by LTO in the form of -Waggressive-loop-optimizations.

Closes: https://bugs.gentoo.org/875530
Closes: https://bugs.gentoo.org/881461
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>