app-emulation/qemu: Update live ebuild
Because we want to download all dependencies in fetch phase
(fixed in [1]), we keep a list of subprojects among with their
latest commit hash. Well, in commit [2] qemu bumped
'berkeley-testfloat-3'. Reflect that in our live ebuild.
Unfortunately, I've missed this in qemu-8.2.0 timeframe (in which
the qemu commit was merged), but there's no real harm as these
subprojects are part of dist tar. IOW, it's only live ebuild that
needs to fetch subprojects explicitly.

Also, some patches don't apply cleanly anymore. Rebase them.

Oh, and 'nios2' target was removed [3].

1: https://github.com/gentoo/gentoo/pull/32684
2: https://gitlab.com/qemu-project/qemu/-/commit/c01196bdddc280ae3710912e98e78f3103155eaf
3: https://gitlab.com/qemu-project/qemu/-/commit/6c3014858c4c0024dd0560f08a6eda0f92f658d6

Signed-off-by: Michal Privoznik <michal.privoznik@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>