app-dicts/myspell-en: add 20240601, remove redundant l10n_en USE flag
In the previous versions of app-dicts/myspell-en there was a
IUSE=l10n_en, which unconditionally enabled all the English varieties.
This USE flag is redundant, since the purpose it serves can be as easily
achieved by switching the individual USE flags on. Hence, the
IUSE=l10n_en is removed.
All other English varieties' USE flags are set to be enabled by default,
to avoid the need to manually set them in order to mimic the old
behaviour (which might have been required in the most cases).

Additionally, as per the dictionary specification in the
./dictionaries.xcu file in the root of the distfile, American thesaurus
is suitable for any English variety; British hyphenation is for any
non-American English variety. The ebuild is changed to reflect this.

Closes: https://bugs.gentoo.org/481618
Signed-off-by: Zurab Kvachadze <zurabid2016@gmail.com>