app-dicts/myspell-en: add 20240601, remove redundant l10n_en USE flag
In the previous versions of app-dicts/myspell-en there was a
IUSE=l10n_en, which unconditionally enabled all the English varieties.
This USE flag is redundant, since the purpose it serves can be as easily
achieved by switching the individual USE flags on. Hence, the
IUSE=l10n_en is removed.
All other English varieties' USE flags are set to be enabled by default,
to avoid the need to manually set them in order to mimic the old
behaviour (which might have been required in the most cases).

Closes: https://bugs.gentoo.org/481618
Closes: https://github.com/gentoo/gentoo/pull/36552
Signed-off-by: Zurab Kvachadze <zurabid2016@gmail.com>
Signed-off-by: Yixun Lan <dlan@gentoo.org>