media-libs/liblo: replace useless sed with configure option
 * SED: the following did not cause any changes
 *     sed -i '/^SUBDIRS =/s/examples//' Makefile.am || die;
 * no-op: -e /^SUBDIRS =/s/examples//

In 2014, the Makefile.am was reworked (in version 0.29):
https://sourceforge.net/p/liblo/git/ci/02f7754e245080626a590e43c190dfd4abe0d72f/

This was a side effect of making it into a configure option, which means
patching is no longer needed.

As a result, the intended goal of this sed hasn't worked for a decade --
the examples were compiled by portage and then discarded as they aren't
installed. It is "just" wasted work. But we can avoid that...

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>