dev-games/libnw: remove highly confusing sed to "respect CC"
But it's a generated automake Makefile.in, and the sed is removing every
line containing "CC = @CC@", so it's puzzling that anyone ever thought
to do this.

It was originally added in gentoo-historical-2:

  29 Jan 2009; Michael Sterrett <mr_bones_@gentoo.org> libnw-1.30.02.ebuild:
  EAPI=2; respect CC

And 15 years later it's the same (2003) tarball and the sed is decidedly
not needed (nor ever was).

Remove this bloaty junk.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>