sys-apps/pcsc-lite: enable py3.12
I don't really use nor know this package, but was blocking
testing other packages on a 3.12-only system.

At a glance pcsc-spy for which this uses python for does
not seem inheritently broken, nor is there pending 3.12
fixes upstream for it that I can see. So sounds safe.

Referencing bug #927448 as well given it requested 3.12,
not that I'm going to look at bumping this myself.

Bug: https://bugs.gentoo.org/927448
Closes: https://bugs.gentoo.org/929819
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>