linux-mod-r1.eclass: update depmod --config paths for kmod-32
kmod-32's tools/depmod.c adds 'DISTCONFDIR "/depmod.d"' which
is $prefix/lib/depmod.d.

Largely a no-op on merged-usr because of the /lib -> /usr/lib symlink,
but could potentially impact split-usr if both directories get used
(albeit this is only used when a different ROOT is used).

Ordering matches the current depmod.c (etc first, then run, etc...).

On a side-note, kmod-32 also now has configurable /lib/modules path
albeit this is not done on Gentoo at the moment (the default without
the prefix is used). May need updates in the future to figure out
which to use though.

Don't believe this is worth a ML review.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>