dev-debug/dtrace: new package, add 2.0.1.1, add 9999
This is the modern incarnation of DTrace. It is now BPF-based and requires
no out-of-tree code or modules.

i.e. it Just Works with sys-kernel/gentoo-kernel[-hardened,debug] (*).

Please see https://wiki.gentoo.org/wiki/DTrace and report any issues.

There is still some pending work to be done for integration in Gentoo:
* Figuring out collision with dev-debug/systemtap
* Wiring up DTrace/systemtap probe support in more packages
* Deciding on consistent naming/global USE for that support

But the tooling works as-is. Enjoy!

(*) -hardened is needed for now because of CONFIG_GCC_PLUGIN_RANDSTRUCT,
see https://gcc.gnu.org/PR84052. Investigation is ongoing on fixing this.

Signed-off-by: Sam James <sam@gentoo.org>