sys-devel/bpf-toolchain: new package, add 14.2.0
Heavily inspired by, and based on, dev-util/mingw64-toolchain. This provides
the bpf-unknown-none toolchain for use with DTrace (to be packaged) and
systemd.

Users may wish to use crossdev if they want to customise their toolchain -
this is "just" a convenience option for e.g. aforementioned ebuilds to depend
on and have things Just Work.

TODO:
* mingw64-toolchain has bin-symlinks custom-cflags debug +strip, do we
  want/need any of those?

Bug: https://bugs.gentoo.org/917228
Signed-off-by: Sam James <sam@gentoo.org>