x11-libs/cmrt: mark as LTO-unsafe, strict-aliasing unsafe
The software is dead upstream. In August 2022 (almost immediately after
the package was added to the tree), the README.md was updated to say:

> Intel has ceased development and contributions including, but not
> limited to, maintenance, bug fixes, new releases, or updates, to this
> project.
>
> Intel no longer accepts patches to this project.

and the github repository was marked as archived.

No point in reporting anything, because you can't even if you want to.
Just mark it as unsafe and move on.

Closes: https://bugs.gentoo.org/864409
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>