dev-qt/qtwebengine: improve clang-18 workaround w/ -mevex512 (qt6)
Hoping it will be a short-lived and that this will be improved/fixed
in clang itself.

(have not tried nor looked at qtwebengine:5)

For some rough explanation from the little I get from this:

clang-18 added -mevex512 (missing from 17), and then -march=native
is a bit quirky in that unlike -march=exact it goes out of its way
to disable it resulting in e.g.

	-march=skylake -mavx512f = -mevex512 is auto-enabled
	-march=skylake -mevex512 = not "enabled" but can be used
	-march=native(skylake) -mavx512f = forced off(!)

And then units that use avx512 / pass -mavx512f (for use with runtime
cpu detection) end in build failure without evex512.

Always passing -mevex512 on a machine without avx512 "seems" safe,
it does not even set __EVEX512__ and believe won't use any avx512
instructions on a whim (__EVEX512__ does get set if add -mavx512f).
Or at least my skylake (not skylake-x) passes test + can use the
qtwebengine built that way.

Considered passing only for files that need it at first with a patch
(sounded safer), but chromium's Gn files don't have a variable to test
clang version that I could see (or at least not in old qtwebengine) and
didn't want this to become more involved nor use conditional patching.

The !avx512 check may not be super necessary, but have not dug into
the implications of forcing it when avx512 is actually enabled (sounds
there are cases where it needs to be off, leaving it to compiler).

Bug: https://bugs.gentoo.org/931623
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>