dev-lang/perl: drop filter-lto
This was added in f8f7c133033918b061a2745e0ec79c25f7e04df5 (thx to csfore
for finding it), but tests pass for me with 5.38.2 with this removed
and with -O3 -flto -march=znver2.

I see indications other distros are building Perl with LTO and some
(old) bugs upstream where it appears supported.

Signed-off-by: Sam James <sam@gentoo.org>