Fix the GRUB configuration script for memtest86+ EFI builds
The current GRUB configuration script for memtest86+ doesn't produce
working entries as it tries to load the executable images using the
`chainloader` command, this fails to load the images.

The executables produced by the build appear to be using the Linux EFI
boot stub and thus can be loaded using the `linux` command instead. This
patch does just that.

Closes: https://bugs.gentoo.org/931825
Signed-off-by: Gabriele Svelto <gsvelto@mozilla.com>