zig.eclass: new eclass
This eclass is built on top of `zig-utils.eclass` and works with
`zig build` sub-commands and `build.zig` system, here called ZBS or
Zig Build System. By default it adds convenient default dependencies
and phase functions (which can be opt-out by setting `ZIG_OPTIONAL=1`)
that:
 * work with dependencies from official package manager,
 * configures basic sane options for `zig build`, which can be further
   complemented by ebuild authors and users,
 * respects user settings for verbosity and jobs number,
 * has some basic support for cross-compilation, which works on my PC
   with `crossdev` and QEMU-ed Docker container, but it was not tested
   on real machine yet.

This is a major step in unifying behaviour and de-duplicating logic
spanning multiple ebuilds (target convertion and jobs amount
in `dev-lang/zig`, different Zig search and build options configuration
in `sys-fs/ncdu` and other ebuilds in community overlays). Mentioned
ebuilds are converted in next commits.

It is recommended to use https://github.com/BratishkaErik/zig-ebuilder
with this eclass to generate base for new ebuilds.

Signed-off-by: Eric Joldasov <bratishkaerik@landless-city.net>