Include also "client-server" section in MySQL/MariaDB init scripts.
For users customizing their MariaDB config files, there's [client-server] section available
(https://mariadb.com/kb/en/configuring-mariadb-with-option-files/) which is encouraged
to hold socket option used in this init script.
MySQL doesn't seem to support it but we already look at [mariadb] section as well. I take it
that extraneous sections are OK to be included here.

Signed-off-by: Emanuel Komínek <drake127@planescape.cz>