linux-mod-r1.eclass: update depmod.d comment
Likely a non-issue with dracut now (untested) but will leave it
alone for now and just updating comment. Maybe if users start
having more problems that'd require it or if we want to more
liberally rename modules install paths while ensuring it does not
use the old one on a same kernel (point of this was just to make
things more robust and is not hard-required).

Also adjust the commented out function to match dracut.conf.d (guess
glob makes it more straight forward, not that it's necessary).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>