dev-db/redis: add 7.2.5
Dependencies were corrected. COMMON_DEPEND was renamed to DEPEND, those
deps are no longer part of BDEPEND as they are not needed in CBUILD.
Test deps were moved to BDEPEND because the testing framework is written
in tcl. The acct-* deps are additionally copied to BDEPEND as they are
required in src_install phase.

Signed-off-by: Petr Vaněk <arkamar@gentoo.org>