dev-vcs/cvs: fix all kinds of Modern C issues
This package release is an extra-vintage 2005 with extra gnulib on top.

Among numerous other issues, the broken autoconf standard macros managed
to break the STDC_HEADERS check, resulting in cvs thinking that there
are no available standard headers, and not bothering to include them.
Ouch!

To fix this, we:
- fix a broken AM_GNU_GETTEXT_VERSION macro to stop autopoint erroring
- rerun eautoreconf, to fix various standard macros with c99 fixes
- add some m4 patches that partially backport gnulib fixes (I think hope
  and pray, since I stopped looking at them after a bit due to the
  inability to backport)

Also, while we are at it, throw in a broken shell quoting fix.

Closes: https://bugs.gentoo.org/898782
Closes: https://bugs.gentoo.org/882225
Closes: https://bugs.gentoo.org/870835
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>