net-misc/dhcpcd: add whitelist for false positive configure implicit decls
arc4random is actually added to glibc now. The checks work there, and
still fail on musl, which they were always supposed to -- the checks are
run with -Werror, and even without that, the linker dies with undefined
references here.

So this is "safe", albeit kinda gross when autoconf exists.

Closes: https://bugs.gentoo.org/924825
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>